refactor: put db to modules
