+++ src/modules/db/db.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { DbService } from 'src/db/db.service';
+import { DbService } from 'src/modules/db/db.service';
 
 @Module({
   exports: [DbService],
